requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
jinja2==3.1.3
pytest==8.0.0
httpx==0.26.0
websockets==12.0
aiofiles==23.2.1
pydantic==2.6.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
bcrypt==4.1.2
pytest-asyncio==0.23.5
sqlalchemy==2.0.27
itsdangerous==2.1.2
python-multipart==0.0.9